Backporting 4.5

+++ ld64/passes/branch_island.cpp
@@ -42,7 +42,8 @@ namespace branch_island {
 
 
 
-struct TargetAndOffset { const ld::Atom* atom; uint64_t offset; };
+
+struct TargetAndOffset { const ld::Atom* atom; uint32_t offset; };
 class TargetAndOffsetComparor
 {
 public:
